Note some connections may be lost when removing a process
The README (and the associated slides) presents this option as a magic bullet for high availability. However, queued connections that have not been accepted will be lost when stopping the "old" process.

There is unfortunately no easy way around as kernel maintainers didn't want to provide a simple option for that (like removing the `SO_REUSEPORT` option). It is believed that it could be done with BPF, but AFAIK nobody came up with actual code. The only solution for a seamless reload is to give the old socket to another instance.